make read handler parent of update, create and delete
